feat: Refactor UserContext structure and update imports across components
